refactor: move gale specific configs into gale pkg
