+++ pkg/gale/ghx.go
@@ -1,4 +1,4 @@
-package tools
+package gale
 
 import (
 	"context"
